Move transaction validator to its own package
